requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
biopython==1.81
biotite==0.37.0
certifi==2022.12.7
charset-normalizer==2.1.1
cmake==3.25.0
einops==0.6.1
fair-esm==2.0.0
filelock==3.9.0
frozenlist==1.3.3
fsspec==2023.6.0
-e git+https://github.com/Graylab/GeoDock.git@e7d3abc1dfab23d5e553efa63ca9020921227e19#egg=GeoDock
idna==3.4
Jinja2==3.1.2
joblib==1.3.1
lightning-utilities==0.9.0
lit==15.0.7
MarkupSafe==2.1.2
mpmath==1.2.1
msgpack==1.0.5
multidict==6.0.4
networkx==3.0
numpy # @ file:///home/conda/feedstock_root/build_artifacts/numpy_1687056313536/work
OpenMM==7.7.0
packaging==23.1
# pdbfixer @ file:///home/conda/feedstock_root/build_artifacts/pdbfixer_1631833746479/work
Pillow==9.3.0
protobuf==4.23.3
psutil==5.9.5
pyg-lib==0.2.0+pt20cu118
pyparsing==3.1.0
pytorch-lightning==1.8.6
PyYAML==6.0
requests==2.28.1
scikit-learn==1.3.0
scipy==1.11.1
sympy==1.11.1
tensorboardX==2.6.1
threadpoolctl==3.1.0
torch==2.0.1+cu118
torch-cluster==1.6.1+pt20cu118
torch-geometric==2.3.1
torch-scatter==2.1.1+pt20cu118
torch-sparse==0.6.17+pt20cu118
torch-spline-conv==1.2.2+pt20cu118
torchaudio==2.0.2+cu118
torchmetrics==0.11.4
torchvision==0.15.2+cu118
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.4.0
urllib3==1.26.13
yarl==1.9.2
